+++ src/aras.jl
@@ -153,4 +153,7 @@ function aras(setting::MCDMSetting)::ARASResult
     )
 end
 
+function aras(mat::Matrix, weights::Array{Float64, 1}, fns::Array{Function, 1})::ARASResult
+	return aras(makeDecisionMatrix(mat), weights, fns)
+end
 end # end of module ARAS